+++ app/src/main/java/com/example/trip/models/Availability.kt
@@ -2,4 +2,4 @@ package com.example.trip.models
 
 import java.time.LocalDate
 
-data class Availability (val id: Int, val userId: Int, val startDate: LocalDate, val endDate: LocalDate)
+data class Availability (val id: Long, val userId: Long, val startDate: LocalDate, val endDate: LocalDate)
