+++ app/src/main/java/com/example/trip/usecases/accommodation/PostAccommodationUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 
 class PostAccommodationUseCase @Inject constructor(private val accommodationsRepository: AccommodationsRepository) {
 
-    suspend operator fun invoke(groupId: Int, accommodationBase: Pair<String, String?>): Resource<Unit> {
+    suspend operator fun invoke(groupId: Long, accommodationBase: Pair<String, String?>): Resource<Unit> {
         return accommodationsRepository.postAccommodation(groupId, accommodationBase)
     }
